Sequencer: Use a custom interface and tick method for sequene actors.
This is because level sequences tick at a specific point in time during the game loop. Adding a Tick override makes the BP compiler enable ticking on the actor, which means that it will tick twice (once with the custom sequencer tick manager, and once with the normal actor tick groups). This will not only make the sequence go twice as fast, it will also cause other problems around starting/stopping. So here we make the sequence tick manager use a custom tick method, which frees up the standard tick to do other things.

Note that we need to introduce an interface for this custom tick method because the tick manager, which resides in the core MovieScene module, doesn't know anything about sequence actors such as LevelSequenceActor.

#jira UE-114460
#rb max.chen

#ROBOMERGE-OWNER: ludovic.chabant
#ROBOMERGE-AUTHOR: ludovic.chabant
#ROBOMERGE-SOURCE: CL 16302522 via CL 16303238 via CL 16362240
#ROBOMERGE-BOT: RELEASE (Release-4.27 -> Main) (v804-16311228)
#ROBOMERGE[bot4]: -BOT4-Release-Engine-Staging

[CL 16362267 by ludovic chabant in Main branch]