Reduce groom asset size.
This CL splits attribute data into two Attribute0 and Attribute1 , with Attribute1 being optional,
GroomAsset Attribute1 and Material data are now optionnaly serialize, loaded, and added as resources. This is based if the hair description contains or not such a data.

This allows to save up to 8bytes per hair vertex. This reduces the GroomAsset footprint by ~18%

#rb none
#jira FROST-2658

[CL 16369361 by Charles deRousiers in ue5-main branch]