Reduce groom asset size.
This CL splits attribute data into two Attribute0 and Attribute1 , with Attribute1 being optional,
GroomAsset Attribute1 and Material data are now optionnaly serialize, loaded, and added as resources. This is based if the hair description contains or not such a data.

This allows to save up to 8bytes per hair vertex. This reduces the GroomAsset footprint by ~18%

#rb none
#jira FROST-2658

#ROBOMERGE-SOURCE: CL 16369361 in //UE5/Main/...
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v804-16311228)

[CL 16369400 by charles derousiers in ue5-release-engine-test branch]