#Jira UE-114896
A process has been added after the import to avoid faraway points. This process was based on the aabb of the topological vertices of the body. In this case, the aabb of the vertices is very small compare to the body. This process make more trouble than it fixes problems. So it has been removed.

#rb jeanluc.corenthin

#ROBOMERGE-SOURCE: CL 16392885 in //UE4/Release-4.27/...
#ROBOMERGE-BOT: RELEASE (Release-4.27 -> Main) (v804-16311228)

[CL 16392915 by david lesage in Main branch]