UdpMessaging: simplify error delegate handling for processor errors
This change addresses a crash that can potentially occur if a message
processing error occurs near the time the UdpMessaging module is
unloaded (typically at engine shutdown).

The crash was happening because the transport's TransportErrorDelegate
was being bound with BindRaw() to a function on the module, but a copy of
the delegate was being created by a lambda capture and scheduled to run
later on the main thread via the TaskGraph. Even though the original delegate
was being unbound at exit, the copy captured in the lambda remained bound
and could execute sometime after the module was unloaded and the UObject
system was no longer available. Trying to access the settings CDO at that point
would lead to a crash.

It turns out that the processor error does not actually need to propagate all the
way up to the module. The AutoRepair routine that is launched in response to an
error does not use any state from the module itself, and in fact is only intended to
keep trying as long as the transport remains valid. We can instead have the
transport manage the AutoRepair routine and cut out a level of delegate
invocation, and the transport no longer needs to expose a delegate publicly.

#jira UE-104952
#rb francis.hurteau

#ROBOMERGE-SOURCE: CL 16393910 in //UE4/Release-4.27/... via CL 16393951
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Staging) (v804-16311228)

[CL 16393972 by matt johnson in ue5-release-engine-staging branch]