Changed UV error weights to be UV area normalized. This makes the most sense to me for how to see UV error as texel error in screen space. Although I feel it is theoretically better this doesn't appear to have any noticable impact, likely because UV error has so far been set vanishingly low.
UV mirror seams are now considered an attribute discontinuity. This is handled through a cheap trick of changing the tops bits of the material index per triangle.

This should not strictly be necessary because the QEM could account for the difference but to get it to properly catch issues in practice required increasing the UV error weight high enough that it impacts all meshes including ones that didn't have noticable issues. Targeting the mirror case specifically was better even though it shouldn't be special theoretically.

This fixes Nanite UV bugs in Frosty.
#jira UE-113150

#ROBOMERGE-SOURCE: CL 16411470 in //UE5/Private-Frosty/...
#ROBOMERGE-BOT: STARSHIP (Private-Frosty -> Main) (v804-16311228)

[CL 16412628 by brian karis in ue5-main branch]