Change hair strands data to be lazily loaded to avoid loading all data while only a portion will be used.
This CL changes the data to serialize into byte bulk data, and for its content to not be inlined, meaning it need to be explicitly loaded from the archive to acces its content. This works only for cook build. With editor data, the data are fetched and loaded from DDC as usual.

#rb none
#jira FROST-2658

#ROBOMERGE-SOURCE: CL 16435448 in //UE5/Private-Frosty/...
#ROBOMERGE-BOT: STARSHIP (Private-Frosty -> Main) (v814-16434418)

[CL 16435454 by charles derousiers in ue5-main branch]