#Jira UE-114896
A process has been added after the import to avoid faraway points. This process was based on the aabb of the topological vertices of the body. In this case, the aabb of the vertices is very small compare to the body. This process make more trouble than it fixes problems. So it has been removed.

#rb jeanluc.corenthin

#ROBOMERGE-OWNER: david.lesage
#ROBOMERGE-AUTHOR: david.lesage
#ROBOMERGE-SOURCE: CL 16392885 in //UE4/Release-4.27/... via CL 16392915 via CL 16447260 via CL 16447533
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Test -> Release-Engine-Staging) (v818-16446889)

[CL 16447561 by david lesage in ue5-release-engine-staging branch]