Copied from CL 15016028:
Fixed an import problem with FBXs containing LODs with meshes which can't be triangulated (e.g. due to being non-manifold).
Previously, an error in an individual FBX mesh would cause the import for the current LOD to stop but not fail, resulting in an incomplete mesh. Now, problematic meshes will be omitted and a warning generated. If a LOD cannot be imported at all, it will be auto-generated instead.
#jira UE-107915
#rb Alexis.Matte

[CL 16483307 by Richard TalbotWatkin in 4.27 branch]