[Insights] Cache layer
Introduces a persistent storage system for Unreal Insights. The cache is based on fixed block sizes. Users give a unique name which is associated with a unique caching index. User can then allocate fixed sized blocks of memory and store data inside. The system make no assumption of the saved data. Each block is ref counted and as it is released the memory is committed to disk.

Additionally a small user data section is provided per cache index for persistent storage of data which does not fit neatly into the block model, such as element counts.

#rb ionut.matasaru

[CL 16501017 by Johan Berg in ue5-main branch]