[Insights] Persistent symbol storage for module provider
Implements persistent symbol storage using the cache layer. Symbols are stored in a cached paged array and with references to a cached string store for module and symbol names and filenames. Previously saved symbols are loaded at the start of the session and any new symbols that are discovered during the session is stored on close.

#rb ionut.matasaru

[CL 16512568 by Johan Berg in ue5-main branch]