Fix issue with flickering in Vulkan.
From the moment a buffer gets locked/unlocked, it modifies the internal Current.Size of that buffer.  That Current.Size is then used as the value to represent the "entire size" of that buffer for clears/fills/binds.  The problem is when a buffer gets reused (rdg's buffer pool, for example):  if it was used in a code path that locks it, then used in a code path that never locks it and expects to use the full size, then the buffer is limited to the size it had in its previous life.  It is also an issue if we lock a buffer at a partial size, but then expect it to be bound at its full size (to read from the filled portions and write to other parts).

NOTE: Fill commands still go beyond used size to clear the entire buffer, but it did before...  Will be tweaked in a later optimization.

#rb rolando.caloca
#jira UE-117547
[FYI] mihnea.balta,dmitriy.dyomin

#ROBOMERGE-SOURCE: CL 16643730 in //UE5/Main/...
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v833-16641396)

[CL 16643733 by jeannoe morissette in ue5-release-engine-test branch]