Implement path compaction for the path tracer to keep GPU lanes more fully occupied as rays bounce around the scene
The basic idea is to breakup the bounce loop in the path tracer into independent launches so we can group active paths together. This step is done with wave ops + atomics. The implementation is rather naive at the moment, but it already gives a 2x speedup on most scenes. This is on par or better than the speedup from the coherent sampler, but without inposing any constraints on the random numbers.
#preflight 60db7bf119a0060001f8e9a9

[CL 16800791 by chris kulla in ue5-main branch]