Fix invalid source state for the first transition on large structured buffers.
FD3D12DynamicRHI::CreateD3D12Buffer assumed static placed buffers will require resource tracking and computed the initial state accordingly. However, is the buffer is read-only, FD3D12Resource::DetermineResourceStates forces tracking to off after the resource is created, and overrides the default state to D3D12_RESOURCE_STATE_GENERIC_READ, even though the resource was created in the state originally computed by CreateD3D12Buffer. This mismatch caused the source state to be incorrect when the buffer was transitioned to COPY_DEST for uploading the initial contents.

The solution is to make CreateD3D12Buffer aware of the fact that read-only resources don't need tracking, in which case it uses GetDefaultInitialResourceState() to determine the initial state, which corresponds to what DetermineResourceStates computes.

#preflight 60f99a432248260001476d5f
#rb Kenzo.TerElst
#rnx

#ROBOMERGE-SOURCE: CL 16953257 via CL 16953264
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v838-16927207)

[CL 16953268 by mihnea balta in ue5-release-engine-test branch]