Added support for editing ISM partition actor instances
This adds a new interface, IISMPartitionInstanceManager, which is the ISM partition actor equivalent of ISMInstanceManager, but works on ISM partition client instances rather than ISMC instances. Internally APartitionActor will forward the underlying ISMC edit request to the appropriate IISMPartitionInstanceManager, if any.

An ISM partition client that wants to provide editing support for its instances must either register a manager via APartitionActor::RegisterClientInstanceManager, or a proxy that can provide a manager via APartitionActor::RegisterClientInstanceManagerProvider. If the manager/provider instance registered is a UObject, then that reference will be serialized and loaded again automatically. If the manager/provider is a normal C++ instance, then the client owner must re-register it manually post-load.

Unlike ISMC instances, which are always editable unless their owner actor disables it, ISM partition actor instances are only editable if a client registers a manager as described above. This is because only the client has enough knowledge to safely edit its instances.

ISM partition actors also required supporting a new editing feature; instance selection groups. This is due to an ISM partition client instance potentially being made up of several ISMC instances that must be treated as a single logical entity for editing. To achieve this we've added an ISMInstanceManager::ForEachSMInstanceInSelectionGroup function that can be overridden by any manager that needs to provide instance grouping, and which is implemented by APartitionActor to group together all the ISMC instances that belong to a given ISM partition client instance.

Our proof-of-concept for ISM partition actor instance editing is foliage, available when the CVar "TypedElements.EnableFoliageInstanceElements" is enabled. This takes the proof-of-concept work that was already done for ISMC foliage and moves it to a common implementation that works on FFoliageInfo, which both FFoliageStaticMesh (for ISMC foliage) and FFoliageISMActor (for ISM partition actor foliage) call into from their respective manager interfaces.

Misc changes:
 - AISMPartitionActor::SelectISMInstances no longer dirties the component, as the selection state is transient and not serialized on the component.
 - Added a PostSerialize and PostLoad function to foliage implementations, to allow fixing up the IFA and other data.
 - Fixed FFoliageISMActor failing to register its delegates post-load.
 - Fixed FFoliageISMActor crashing due to FoliageTypeActor being null post-load.
 - Fixed FFoliageStaticMesh and FFoliageISMActor leaving raw delegates bound post-destruction.
 - Added the ability for instance managers to override the display name and tooltip used for instances in the details panel.

#jira UETOOL-3785
#rb Brooke.Hubert, Patrick.Enfedaque
#preflight 61016d790bf6fc0001313bf7

#ROBOMERGE-SOURCE: CL 17057600 in //UE5/Main/...
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v850-17047176)

[CL 17058375 by jamie dale in ue5-release-engine-test branch]