UE5.Main:  Fix for color picker closing immediately when environment light mixer window is open.
Tick function for environment light mixer was calling "SDetailsView::SetObject" with a null Object selection every frame, which internally closes any active color pickers.  Now, it only calls that function if the details view in question doesn't already have an empty Object selection.  This was the narrowest change that fixes the bug.  The code that automatically closes color pickers was added to SDetailsView.cpp in CL 16779751, to fix UE-118530, a Pri 1 bug.  There may be an alternate fix that involves refactoring that code, but this seems adequate, and less risky.

#jira UE-121571
#rb daren.cheng
#rnx

#ROBOMERGE-SOURCE: CL 17232975 in //UE5/Main/...
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v855-17104924)

[CL 17232993 by jason hoerner in ue5-release-engine-test branch]