Alembic: update Linux build after ThirdParty move
This includes a change to configure the build to write files to a "deploy" directory for
consistency with the Mac and Windows builds. After building, libraries should then be
copied from there into the "AlembicDeploy" directory, which is where AlembicLib.Build.cs
looks for them.

Alembic expects to find the HDF5 library with a "lib" prefix, and the HDF5 build script
was already configured correctly for this. The existing library file was renamed to
include the "lib" prefix (and the "-static" suffix, which the build also appends), and the
AlembicLib.Build.cs was updated to look for the library file with this name.

#rb anousack.kitisa

[CL 17392559 by matt johnson in ue5-main branch]