Alembic: update Windows build after ThirdParty move and use Visual Studio 2019
We continue to output the build products to a "2015" directory, since
WindowsTargetRules.GetVisualStudioCompilerVersionName() will still report that
value for Visual Studio 2017, 2019, and 2022, as they should all be backwards
compatible.

This change also configures the build to write files to a "deploy" directory for consistency
with the Linux and Mac builds. After building, header files and libraries should then be
copied from there into the "AlembicDeploy" directory, which is where AlembicLib.Build.cs
looks for them. The file and directory paths in all batch scripts were also changed to use
Windows-style paths with backslashes.

#rb anousack.kitisa
#preflight 612d17a1e68f05000143b702

[CL 17392644 by matt johnson in ue5-main branch]