Change groom LOD to have one of the following mode: immediated/predicted/forced * Immediate: LOD is selected on the rendering thread prior to rendering * Predictied: LOD is estimated on the rendering-thread and feedback to the game-thread. The game-thread selects which LOD is appropriate based on the feedback. * Forced: LOD is set by high-level logic on the game-thread, forcing ar particular LOD.
When a groom as one of its LOD/Group with simulation enabled, LOD selection is forced to be predicted. This is mandatory as the simulation needs to know on the game-thread, when to initialized/enabled Niagara components.

This allows to fix case where the simulation is not used/initialized due to starting at a lower LOD (with no sim) and then moving to a higher LOD (with sim), and having the sim. not being kicked-in/initialized properly.

#rb michael.forot
#preflight 6135de6cd031a40001663820

#ROBOMERGE-AUTHOR: charles.derousiers
#ROBOMERGE-SOURCE: CL 17437376 via CL 17439784
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v870-17433530)

[CL 17439930 by charles derousiers in ue5-release-engine-test branch]