Implement BxDF energy conservation for the path tracer
Implement basic energy conservation using albedo tables of the GGX and cloth lobes. Reflection and refraction are tabulated independently.

Also provide analytic approximations for the reflection case if a tabulation is not convenient. If a good analytic approximation for the glass case can be found (requires a 3D fit), we should remove the tabulated implementation which was done mainly to provide a baseline.

The albedo table is also used to improve lobe selection probabilities, which improves sampling of the clearcoat case in particular.

Refactored how the path tracer's internal view state data is stored to keep the structure details fully hidden inside PathTracing.cpp

#rb Patrick.Kelly
#preflight 613814072d09b900014699c9

#ROBOMERGE-AUTHOR: chris.kulla
#ROBOMERGE-SOURCE: CL 17452976 in //UE5/Main/...
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v870-17433530)

[CL 17452996 by chris kulla in ue5-release-engine-test branch]