Horde: Add support for multivalue properties in conditions (such as the 'pool' property on an agent).
When expressions are evaluated, any binary operator is taken to match if any of its values match. For example, a condition matching Pool='Win-UE5' will match any agent in the Win-UE5 pool, even if it also happens to be in another pool. Likewise, Pool != 'Win-UE5' will match any agent in the Win-UE5 pool, even if it also belongs to other pools.

As part of this change, the query syntax has been adjusted to distinguish conditional expressions from comparison expressions. It is no longer possible to write expressions such as "(a == 1) == (b == 2)"; the more expressive form is required ("(a == 1 && b == 2) || (a != 1 && b != 2)".

#ROBOMERGE-AUTHOR: ben.marsh
#ROBOMERGE-SOURCE: CL 17481506 in //UE5/Main/...
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v870-17433530)

[CL 17481511 by ben marsh in ue5-release-engine-test branch]