Transition vertex buffers right before calling IASetVertexBuffers, instead doing it when the high level calls RHISetStreamSource.
If a command list flush happens in the middle of a sequence of draws which use the same vertex buffer, the tracked internal state of the buffer is reset. Since the high level doesn't call RHISetStreamSource again after the flush, no pending barrier was inserted. If a subsequent dispatch needed the buffer in a different state, that transition was erroneously inserted in the patch-up command list, which modified the state too soon (before the draws which still need the buffer in the VB state). Adding the transition in FD3D12DescriptorCache::SetVertexBuffers ensures that the correct state is applied, since this method is called internally before each draw.

#jira FROST-4979
#rb Kenzo.TerElst
#rnx
#preflight 613a08d7b904830001bf5a12
#lockdown Michal.Valient

#ushell-cherrypick of 17489321 by mihnea.balta

[CL 17489558 by mihnea balta in ue5-main branch]