Implement absorption support for refractive translucent materials in the path tracer
The absorption coefficient is derived from the BaseColor of the material and a global scale factor that gives the inverse distance at which the target BaseColor is achieved. Setting this cvar to 0 disables the effect. The default value of 0.01 corresponds to specifing the target color after travelling through 1 meter of surface. This scale factor will not be needed in future material models that provide more direct control over the absorption color.

The implementation remembers the current absorption value as part of the PathState. For shadow rays, when approximate caustics are being used, the contribution is tracked by simply counting up/down at each interface. An offset is used to account for the possibility of light starting inside an object for finite light sources.

Refactored TraceTransparentRay to directly use the PathState struct since it was accessing most of its fields anyway and the function signature was getting too long

#rb Patrick.Kelly
#preflight 61422aac3c7c67000197ca0b

[CL 17524226 by chris kulla in ue5-main branch]