USD: Track ACineCameraActor root components on export and use them as such during import
#jira UE-120826
#rb none

The problem is that ACineCameraActors don't have the actual UCineCameraComponent as the root component: They just have a dummy USceneComponent and a child UCineCameraComponent attached to it.

When exporting to USD we write one prim per component so while everything was technically correct, the issue on the JIRA is that it becomes a bit awkward when you export to USD and import back: The Xform prim becomes just a regular scene component, that's likely going to be collapsed into something else (see on the JIRA how it gets collapsed into the Root actor), and the child Camera prim gets its own ACineCameraActor, with *another* brand new root scene component.

It gets confusing when looking at the USD file too because the parent Xform actor is the one that ends up with the actual original ACineCameraActor transform, and the Camera prim basically is left at identity. This becomes more annoying if we have transform tracks, for example.

I was trying to fix this with some kind of component/prim surgery during export, to basically pretend the CineCameraComponent is the root component, but this can't work: What if the user has some relevant data/animation on that dummy root scene component? It will be lost. It may be used for example to setup some kind of camera pivot point. Also, the user could have placed many other components on that CineCameraActor, or derived its class and reshuffled the component order, etc. Not to mention that handling the transform operations for this hack would be a permanent bug generator because we must compensate camera transforms when writing/reading from USD due to the different camera facing direction...

Instead, the approach on this fix is that whenever we export ACineCameraActors, the Xform prim that corresponds to the root component will receive a new custom attribute that indicates the name of the child prim that corresponds to the actor's UCineCameraComponent, if any. It will still be a perfectly valid setup either way, but if we see that pattern when importing, we'll spawn an ACineCameraActor for the parent Xform prim, and when translating the child Camera prim we'll actually reach out and reuse the parent ACineCameraActor's UCineCameraComponent instead, hopefully recreating the original ACineCameraActor completely.

I had the attribute actually point at a child prim name instead of just saying "hey I'm a cine camera actor" because we could potentially have multiple camera components on that actor, and that way we know for sure we can reuse the UCineCameraComponent when we're parsing that child Camera prim.

The only casualty is that we have to tweak these locations where we try to retrieve the component from a prim twin, as now there are basically two cine camera component cases... I think this is way better than modifying the prim structure though: The two big tasks related to level sequences will need very minor updates (like the one done on the USDStageActor.cpp here, for example).

I've done this targetting 4.27.1, but let me know if you guys think this is too complex for it, and I can just move it to 5.0 instead.

[CL 17570046 by daniel coelho in 4.27 branch]