Fix for some FN tests failing with UE5
There's a UE5 change that results in ScanPathsSynchronous accessing an invalid array index if empty paths are passed in. This could be considered invalid, but delegates are (I think) fired by the asset registry so just early-out'ing with no action might be problematic. As a result I modified the code that logs the paths to handle this.

On the other hand I don't see any need for the Fortnite code to perform this scan if a plugin has no mount points when it's unloaded so I've also modified that code to skip it.

#[review]-17487108 @Matt.Peters @Dave.Belanger
#jira nojira
#preflight https://horde.devtools.epicgames.com/job/613d4fd472f6640001bbfb28?label=18

#ushell-cherrypick of 17487147 by Andrew.Grant

[CL 17571541 by Matt Peters in ue5-main branch]