Fix occasional D3D12 crash after allocating a resource via RHICreateTexture2DFromResource and releasing it.
FD3D12DynamicRHI::CreateTextureFromResource() shouldn't call Location.AsStandAlone(), since that adds the allocation to FD3D12Adapter::TrackedAllocationData. However, since the type is then changed to eAliased, the resource will not be removed from TrackedAllocationData when it's released (FD3D12ResourceLocation::ReleaseResource() only calls UpdateStandAloneStats() for eStandAlone resources, as it should). This skews the stats, but more importantly causes an assert if the memory for the released FD3D12ResourceLocation object is reused by a subsequent allocation.

#jira UE-120370
#rnx
#rb Kenzo.TerElst

#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 17608117 in //UE5/Main/...
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v871-17566257)

[CL 17608124 by mihnea balta in ue5-release-engine-test branch]